v4.0 A.4: split webhooks/dispatcher.py per concern
Extract the 1748-line dispatcher into focused per-concern modules.
Outbox contract (slice 13/14), retry state machine (slice 2),
lease semantics (slice 2 round 3), terminal-succeeded trigger
(slice 2 round 24) — all preserved exactly.

- mnemos/webhooks/__init__.py: re-exports public surface
- mnemos/webhooks/dispatcher.py: dispatch() + _dispatch_on_conn (thin)
- mnemos/webhooks/outbox.py: transactional INSERT contract
- mnemos/webhooks/workers.py: lifespan worker loops
- mnemos/webhooks/lease.py: claim / guard / release
- mnemos/webhooks/sender.py: HTTP send pipeline + body capture
- mnemos/webhooks/finalize.py: terminal state transitions
- mnemos/webhooks/chain.py: retry-chain management
- mnemos/webhooks/types.py: dataclasses + send semaphore
- mnemos/webhooks/repair.py: broken-state cleanup
- mnemos/webhooks/_signing.py: HMAC

No state-machine changes. No wire-shape changes. 925 tests pass.

Authored-By: Codex
