+++ mnemos/webhooks/__init__.py
@@ -1,1 +1,13 @@
 """Webhook dispatch infrastructure."""
+from __future__ import annotations
+
+from .dispatcher import _dispatch_on_conn, dispatch
+from .workers import delivery_worker_loop, recovery_worker_loop, repair_worker_loop
+
+__all__ = [
+    "dispatch",
+    "_dispatch_on_conn",
+    "repair_worker_loop",
+    "delivery_worker_loop",
+    "recovery_worker_loop",
+]
