+++ mnemos/webhooks/dispatcher.py
@@ -1,173 +1,27 @@
-"""Webhook dispatcher — records delivery intent, fires HTTP POST, retries on failure.
-
-Usage from handlers:
-    from mnemos.webhooks.dispatcher import dispatch
-    await dispatch("memory.created", {"memory_id": ..., "content": ...}, conn=conn)
-
-Design notes
-------------
-- Delivery is durable via the `webhook_deliveries` table. A row is written
-  before the HTTP call so crashes between queue and send can be replayed.
-- Initial attempt runs inline as a background task (asyncio.create_task via
-  _schedule_background). On failure, a new delivery row is scheduled at the
-  next backoff interval; the failed attempt is marked `abandoned` with
-  `superseded=TRUE` so audit queries can distinguish retry-chain advancement
-  from final-attempt failure. Workers claim due rows
-  by writing a short-lived lease, release the database connection, and only
-  then perform DNS validation and the outbound POST. The lease is the
-  authoritative delivery budget: DNS validation and HTTP response headers run
-  under a wall-clock deadline derived from the lease, leaving a buffer for the
-  finalize transaction before recovery can reclaim the attempt. Once headers are
-  received, the status code is the delivery acknowledgement and the attempt is
-  finalized immediately with response_body=NULL before any body capture or
-  stream cleanup. Response-body capture is best-effort post-finalize audit data
-  and cannot turn a 2xx into a retry.
-- HMAC-SHA256 signature over the raw JSON body bytes. Receivers verify with
-  the per-subscription secret returned once at create time.
-
-Retry schedule: 1 minute, 5 minutes, 30 minutes, 2 hours. After 4 failed
-attempts a delivery is marked 'abandoned'.
-"""
+"""Public webhook dispatch entry points."""
 from __future__ import annotations
 
 import asyncio
-import hashlib
-import hmac
-import json
 import logging
-import os
+import sys
 import time
-import uuid
-from dataclasses import dataclass, field
-from datetime import datetime, timedelta, timezone
-from typing import Any, Awaitable, Dict, Iterable, Optional
+import types as _module_types
+from typing import Any, Dict, Optional
 
 import asyncpg
 import httpx
 
+from . import _signing as webhook_signing
+from . import chain as webhook_chain
+from . import finalize as webhook_finalize
+from . import lease as webhook_lease
+from . import repair as webhook_repair
+from . import sender as webhook_sender
+from . import types as webhook_types
+from . import workers as webhook_workers
+from .outbox import _dispatch_on_conn as _outbox_dispatch_on_conn
+
 logger = logging.getLogger(__name__)
-
-# Retry schedule in seconds: 1m, 5m, 30m, 2h
-BACKOFF_SCHEDULE = [60, 300, 1800, 7200]
-MAX_ATTEMPTS = len(BACKOFF_SCHEDULE)  # = 4
-DNS_TIMEOUT = float(os.getenv("WEBHOOK_DNS_TIMEOUT", "10.0"))
-DELIVERY_TIMEOUT = float(os.getenv("WEBHOOK_HTTP_TIMEOUT", "10.0"))
-WEBHOOK_LEASE_SECONDS = int(os.getenv(
-    "WEBHOOK_LEASE_SECONDS",
-    str(max(90, int(DNS_TIMEOUT + DELIVERY_TIMEOUT + 30))),
-))
-WEBHOOK_FINALIZE_BUFFER_SECONDS = float(os.getenv("WEBHOOK_FINALIZE_BUFFER_SECONDS", "5.0"))
-WEBHOOK_RESPONSE_BODY_MAX_BYTES = int(os.getenv("WEBHOOK_RESPONSE_BODY_MAX_BYTES", "2048"))
-WEBHOOK_RESPONSE_BODY_CAPTURE_TIMEOUT_SECONDS = 5.0
-WEBHOOK_POST_HEADER_CLEANUP_TIMEOUT_SECONDS = float(
-    os.getenv("WEBHOOK_POST_HEADER_CLEANUP_TIMEOUT_SECONDS", "5.0")
-)
-WEBHOOK_MAX_CONCURRENT_SENDS = int(os.getenv("WEBHOOK_MAX_CONCURRENT_SENDS", "64"))
-NEW_CODE_WRITER_REVISION = 1
-NON_IDENTITY_RESPONSE_BODY_PREVIEW_BYTES = 256
-MIN_SEND_WINDOW_SECONDS = 1.0
-RECOVERY_POLL_INTERVAL = 30.0          # seconds between recovery-worker passes
-REPAIR_BURST_SECONDS = float(os.getenv("WEBHOOK_REPAIR_BURST_SECONDS", "60.0"))
-REPAIR_BURST_INTERVAL = float(os.getenv("WEBHOOK_REPAIR_BURST_INTERVAL", "5.0"))
-REPAIR_PERIODIC_INTERVAL = float(os.getenv("WEBHOOK_REPAIR_PERIODIC_INTERVAL", "300.0"))
-TERMINAL_DELIVERY_STATUSES = frozenset((
-    "succeeded",
-    "abandoned",
-))
-LIVE_DELIVERY_STATUSES = frozenset(("pending", "retrying"))
-_send_semaphore: asyncio.Semaphore | None = None
-
-
-def _derive_total_send_deadline_seconds(
-    lease_seconds: int,
-    finalize_buffer_seconds: float,
-) -> float:
-    """Derive the single wall-clock DNS+HTTP budget from the attempt lease."""
-    if finalize_buffer_seconds <= 0:
-        raise ValueError("WEBHOOK_FINALIZE_BUFFER_SECONDS must be positive")
-    deadline = float(lease_seconds) - finalize_buffer_seconds
-    if deadline <= 0:
-        raise ValueError(
-            "WEBHOOK_LEASE_SECONDS must be greater than WEBHOOK_FINALIZE_BUFFER_SECONDS "
-            "so webhook sends leave time for finalization before lease expiry"
-        )
-    return deadline
-
-
-if WEBHOOK_RESPONSE_BODY_MAX_BYTES <= 0:
-    raise ValueError("WEBHOOK_RESPONSE_BODY_MAX_BYTES must be positive")
-if WEBHOOK_POST_HEADER_CLEANUP_TIMEOUT_SECONDS <= 0:
-    raise ValueError("WEBHOOK_POST_HEADER_CLEANUP_TIMEOUT_SECONDS must be positive")
-
-# Keep the lease as the only operator-facing ownership budget. This derived
-# value validates startup configuration; actual sends use the DB-returned claim
-# timestamp pair so an app pause after claim cannot spend a stale full budget.
-TOTAL_SEND_DEADLINE_SECONDS = _derive_total_send_deadline_seconds(
-    WEBHOOK_LEASE_SECONDS,
-    WEBHOOK_FINALIZE_BUFFER_SECONDS,
-)
-
-WEBHOOK_RETRY_SUCCESSOR_REPAIR_SQL = """
-    UPDATE webhook_deliveries d
-    SET status = 'abandoned',
-        superseded = TRUE,
-        status_updated_at = clock_timestamp(),
-        lease_token = NULL,
-        lease_expires_at = NULL
-    WHERE d.status IN ('pending', 'retrying')
-      AND (d.lease_token IS NULL OR d.lease_expires_at < clock_timestamp())
-      AND (
-        EXISTS (
-          SELECT 1
-          FROM webhook_deliveries newer
-          WHERE newer.subscription_id = d.subscription_id
-            AND newer.event_type = d.event_type
-            AND newer.payload_hash = d.payload_hash
-            AND newer.attempt_num > d.attempt_num
-        )
-        OR EXISTS (
-          SELECT 1
-          FROM webhook_deliveries peer
-          WHERE peer.subscription_id = d.subscription_id
-            AND peer.event_type = d.event_type
-            AND peer.payload_hash = d.payload_hash
-            AND peer.status = 'succeeded'
-        )
-      )
-"""
-
-
-@dataclass(frozen=True)
-class _DeliveryResult:
-    succeeded: bool
-    response_status: Optional[int] = None
-    response_body: Optional[str] = None
-    error: Optional[str] = None
-
-
-@dataclass(frozen=True)
-class _LeaseExpiredBeforeSend(_DeliveryResult):
-    """Marker result for a lease window that elapsed before any POST began."""
-
-
-@dataclass(frozen=True)
-class _PostHeaderDeliveryResult(_DeliveryResult):
-    """Status-code result plus open response resources for post-finalize audit work."""
-
-    delivery_id: Any = field(default=None, repr=False, compare=False)
-    response: Any = field(default=None, repr=False, compare=False)
-    stream_cm: Any = field(default=None, repr=False, compare=False)
-    client_cm: Any = field(default=None, repr=False, compare=False)
-
-
-@dataclass(frozen=True)
-class _ClaimedDelivery:
-    delivery: asyncpg.Record
-    lease_token: str
-    pre_claim_monotonic: float
-
-
-# ── Public surface ────────────────────────────────────────────────────────────
 
 
 async def dispatch(
@@ -211,7 +65,7 @@ async def dispatch(
 
     from mnemos.core.lifecycle import _pool as lifecycle_pool  # noqa: WPS433
     if not lifecycle_pool:
-        logger.warning("webhook dispatcher: no DB pool — skipping event %s", resolved_event_type)
+        logger.warning("webhook dispatcher: no DB pool - skipping event %s", resolved_event_type)
         return
 
     async with lifecycle_pool.acquire() as acquired_conn:
@@ -233,1516 +87,144 @@ async def _dispatch_on_conn(
     namespace: Optional[str] = None,
 ) -> None:
     """Insert delivery intents using an already-selected connection."""
-    subs = await _matching_subscriptions(conn, event_type, owner_id, namespace)
-    if not subs:
-        return
-
-    body = json.dumps({
-        "event": event_type,
-        "timestamp": datetime.now(timezone.utc).isoformat(),
-        "data": payload,
-    }, separators=(",", ":"), sort_keys=True)
-    body_hash = hashlib.sha256(body.encode("utf-8")).hexdigest()
-
-    for sub in subs:
-        delivery_id = await conn.fetchval(
-            """
-            INSERT INTO webhook_deliveries
-              (subscription_id, event_type, payload, payload_hash, status, writer_revision)
-            VALUES ($1, $2, $3, $4, 'pending', $5)
-            RETURNING id
-            """,
-            sub["id"], event_type, body, body_hash, NEW_CODE_WRITER_REVISION,
-        )
-        # Schedule the send via the lifecycle-tracked delivery registry
-        # so graceful shutdown lets in-flight attempts finalize. Import lazily to avoid
-        # circular imports at module load time.
-        from mnemos.core.lifecycle import _schedule_delivery_attempt  # noqa: WPS433
-        _schedule_delivery_attempt(_attempt_delivery(str(delivery_id)))
-
-
-async def repair_worker_loop(pool: asyncpg.Pool) -> None:
-    """Background loop: repair superseded retry rows on its own cadence.
-
-    Started from the FastAPI lifespan. Cancels cleanly on shutdown.
-    """
-    logger.info("webhook retry repair worker started")
-    loop = asyncio.get_running_loop()
-    repair_burst_deadline = loop.time() + REPAIR_BURST_SECONDS
-    next_repair_at = loop.time()
-    while True:
-        try:
-            now = loop.time()
-            if now >= next_repair_at:
-                in_burst = now < repair_burst_deadline
-                await _repair_superseded_retrying_deliveries_safely(
-                    pool,
-                    phase="burst" if in_burst else "periodic",
-                )
-                next_repair_at = now + (
-                    REPAIR_BURST_INTERVAL if in_burst else REPAIR_PERIODIC_INTERVAL
-                )
-
-            await asyncio.sleep(max(0.0, next_repair_at - loop.time()))
-        except asyncio.CancelledError:
-            logger.info("webhook retry repair worker cancelled")
-            raise
-        except Exception:  # pragma: no cover — log and keep running
-            logger.exception("webhook retry repair worker iteration failed")
-            await asyncio.sleep(REPAIR_BURST_INTERVAL)
-
-
-async def delivery_worker_loop(pool: asyncpg.Pool) -> None:
-    """Background loop: picks up pending deliveries whose scheduled_at has arrived.
-
-    Started from the FastAPI lifespan. Cancels cleanly on shutdown.
-    """
-    logger.info("webhook delivery recovery worker started")
-    while True:
-        try:
-            await _recover_due_deliveries(pool)
-            await asyncio.sleep(RECOVERY_POLL_INTERVAL)
-        except asyncio.CancelledError:
-            logger.info("webhook delivery recovery worker cancelled")
-            raise
-        except Exception:  # pragma: no cover — log and keep running
-            logger.exception("webhook delivery recovery worker iteration failed")
-            await asyncio.sleep(RECOVERY_POLL_INTERVAL)
-
-
-async def recovery_worker_loop(pool: asyncpg.Pool) -> None:
-    """Compatibility wrapper for the delivery recovery loop."""
-    await delivery_worker_loop(pool)
-
-
-# ── Internals ─────────────────────────────────────────────────────────────────
-
-
-async def _repair_superseded_retrying_deliveries_safely(
-    pool: asyncpg.Pool,
-    *,
-    phase: str,
-) -> None:
-    """Run one repair sweep without killing the recovery worker on failure."""
-    try:
-        result = await repair_superseded_retrying_deliveries(pool)
-        logger.info("webhook retry repair %s sweep result: %s", phase, result)
-    except Exception:  # pragma: no cover — log and keep running
-        logger.exception("webhook retry repair %s sweep failed", phase)
-
-
-async def repair_superseded_retrying_deliveries(pool: asyncpg.Pool) -> str:
-    """Terminalize live attempts made obsolete by successor rows or predecessor success."""
-    async with pool.acquire() as conn:
-        return await conn.execute(WEBHOOK_RETRY_SUCCESSOR_REPAIR_SQL)
-
-
-async def _recover_due_deliveries(pool: asyncpg.Pool, *, limit: int = 50) -> int:
-    """Recover due deliveries by scheduling lifecycle-tracked send attempts."""
-    claim_limit = min(50, limit, _semaphore_available())
-    if claim_limit <= 0:
-        return 0
-
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            claimed_deliveries = await _claim_recoverable_deliveries(conn, limit=claim_limit)
-    from mnemos.core.lifecycle import _schedule_delivery_attempt  # noqa: WPS433
-    for claimed in claimed_deliveries:
-        _schedule_delivery_attempt(
-            _attempt_delivery(str(claimed.delivery["id"]), pool=pool, claimed=claimed)
-        )
-    recovered = len(claimed_deliveries)
-    if recovered:
-        await asyncio.sleep(0)
-    return recovered
-
-
-def _semaphore_available() -> int:
-    """Return the current best-effort free slot count for recovery batch sizing."""
-    return max(0, _get_send_semaphore()._value)
-
-
-async def _claim_recoverable_deliveries(
-    conn: asyncpg.Connection,
-    *,
-    limit: int = 50,
-    lease_seconds: int | None = None,
-) -> list[_ClaimedDelivery]:
-    """Claim due recovery rows before scheduling send tasks."""
-    if limit <= 0:
-        return []
-    if lease_seconds is None:
-        lease_seconds = WEBHOOK_LEASE_SECONDS
-    lease_token = str(uuid.uuid4())
-    pre_claim_monotonic = time.monotonic()
-    rows = await conn.fetch(
-        """
-        WITH claim_clock AS (
-          SELECT clock_timestamp() AS claim_now
-        ),
-        recoverable AS (
-          SELECT d.id
-          FROM webhook_deliveries d, claim_clock
-          WHERE d.scheduled_at <= claim_clock.claim_now
-            AND d.attempt_num <= $1
-            AND d.status NOT IN ('succeeded', 'abandoned')
-            AND NOT d.superseded
-            AND d.status IN ('pending', 'retrying')
-            AND (d.lease_token IS NULL OR d.lease_expires_at < claim_clock.claim_now)
-            AND d.writer_revision = $5
-            AND NOT EXISTS (
-              SELECT 1
-              FROM webhook_deliveries peer
-              WHERE peer.subscription_id = d.subscription_id
-                AND peer.event_type = d.event_type
-                AND peer.payload_hash = d.payload_hash
-                AND peer.status = 'succeeded'
-            )
-            AND (
-              d.status = 'pending'
-              OR (
-                d.status = 'retrying'
-                AND NOT EXISTS (
-                  SELECT 1
-                  FROM webhook_deliveries newer
-                  WHERE newer.subscription_id = d.subscription_id
-                    AND newer.event_type = d.event_type
-                    AND newer.payload_hash = d.payload_hash
-                    AND newer.attempt_num > d.attempt_num
-                )
-              )
-            )
-          ORDER BY d.scheduled_at
-          LIMIT $3
-          FOR UPDATE SKIP LOCKED
-        )
-        UPDATE webhook_deliveries d
-        SET lease_token=$2::uuid,
-            lease_expires_at=claim_clock.claim_now + ($4::int * INTERVAL '1 second'),
-            status=CASE WHEN d.status = 'pending' THEN 'retrying' ELSE d.status END
-        FROM recoverable, webhook_subscriptions s, claim_clock
-        WHERE d.id = recoverable.id
-          AND s.id = d.subscription_id
-        RETURNING d.id, d.subscription_id, d.event_type, d.payload,
-                  d.payload_hash, d.attempt_num, d.status,
-                  d.lease_expires_at, claim_clock.claim_now AS claim_db_now,
-                  $2::uuid AS lease_token,
-                  s.url, s.secret, s.revoked
-        """,
-        MAX_ATTEMPTS,
-        lease_token,
-        limit,
-        lease_seconds,
-        NEW_CODE_WRITER_REVISION,
-    )
-    return [
-        _ClaimedDelivery(
-            delivery=row,
-            lease_token=lease_token,
-            pre_claim_monotonic=pre_claim_monotonic,
-        )
-        for row in rows
-    ]
-
-
-async def _recoverable_delivery_ids(
-    conn: asyncpg.Connection,
-    *,
-    limit: int = 50,
-) -> Iterable[asyncpg.Record]:
-    """Return due delivery rows without claiming them, for diagnostics and tests."""
-    return await conn.fetch(
-        """
-        SELECT d.id FROM webhook_deliveries d
-        WHERE d.scheduled_at <= clock_timestamp()
-          AND d.attempt_num <= $1
-          AND d.status NOT IN ('succeeded', 'abandoned')
-          AND NOT d.superseded
-          AND d.status IN ('pending', 'retrying')
-          AND (d.lease_token IS NULL OR d.lease_expires_at < clock_timestamp())
-          AND d.writer_revision = $3
-          AND NOT EXISTS (
-            SELECT 1
-            FROM webhook_deliveries peer
-            WHERE peer.subscription_id = d.subscription_id
-              AND peer.event_type = d.event_type
-              AND peer.payload_hash = d.payload_hash
-              AND peer.status = 'succeeded'
-          )
-          AND (
-            d.status = 'pending'
-            OR (
-              d.status = 'retrying'
-              AND NOT EXISTS (
-                SELECT 1
-                FROM webhook_deliveries newer
-                WHERE newer.subscription_id = d.subscription_id
-                  AND newer.event_type = d.event_type
-                  AND newer.payload_hash = d.payload_hash
-                  AND newer.attempt_num > d.attempt_num
-              )
-            )
-          )
-        ORDER BY d.scheduled_at
-        LIMIT $2
-        """,
-        MAX_ATTEMPTS, limit, NEW_CODE_WRITER_REVISION,
+    await _outbox_dispatch_on_conn(
+        conn,
+        event_type,
+        payload,
+        owner_id=owner_id,
+        namespace=namespace,
     )
 
-
-async def _matching_subscriptions(
-    conn: asyncpg.Connection,
-    event_type: str,
-    owner_id: Optional[str],
-    namespace: Optional[str],
-) -> Iterable[asyncpg.Record]:
-    """Find non-revoked subscriptions that include this event_type.
-
-    If owner_id/namespace are provided, filter to subscriptions with matching
-    ownership. Otherwise, return all non-revoked matches (useful for
-    system-level events not bound to a caller).
-    """
-    query = """
-        SELECT id, url, events, secret, owner_id, namespace
-        FROM webhook_subscriptions
-        WHERE NOT revoked AND $1 = ANY(events)
-    """
-    args: list = [event_type]
-    if owner_id is not None:
-        query += " AND owner_id = $2"
-        args.append(owner_id)
-        if namespace is not None:
-            query += " AND namespace = $3"
-            args.append(namespace)
-    return await conn.fetch(query, *args)
-
-
-def _get_send_semaphore() -> asyncio.Semaphore:
-    global _send_semaphore
-    if _send_semaphore is None:
-        _send_semaphore = asyncio.Semaphore(max(1, WEBHOOK_MAX_CONCURRENT_SENDS))
-    return _send_semaphore
-
-
-async def _attempt_delivery(
-    delivery_id: str,
-    *,
-    pool: Optional[asyncpg.Pool] = None,
-    claimed: Optional[_ClaimedDelivery] = None,
-) -> bool:
-    """Claim, send, and finalize one delivery without holding DB during I/O."""
-    if pool is None:
-        from mnemos.core.lifecycle import _pool as lifecycle_pool  # noqa: WPS433
-        pool = lifecycle_pool
-    if not pool:
-        logger.warning("webhook dispatcher: no DB pool — skipping delivery %s", delivery_id)
-        return False
-
-    async with _get_send_semaphore():
-        if claimed is None:
-            lease_token = str(uuid.uuid4())
-            claimed = await _claim_delivery(pool, delivery_id, lease_token=lease_token)
-        else:
-            lease_token = claimed.lease_token
-            if not await _guard_preclaimed_delivery_before_send(
-                pool,
-                claimed.delivery,
-                lease_token,
-            ):
-                return False
-        if not claimed:
-            return False
-        result = await _send_claimed_delivery(
-            claimed.delivery,
-            pre_claim_monotonic=claimed.pre_claim_monotonic,
-        )
-        return await _finalize_delivery(pool, claimed.delivery, lease_token, result)
-
-
-async def _claim_delivery(
-    pool: asyncpg.Pool,
-    delivery_id: str,
-    *,
-    lease_token: str,
-    lease_seconds: int | None = None,
-) -> Optional[_ClaimedDelivery]:
-    """Persist a short lease for one due live delivery row."""
-    if lease_seconds is None:
-        lease_seconds = WEBHOOK_LEASE_SECONDS
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            delivery = await _load_delivery_for_claim(conn, delivery_id)
-            if not delivery:
-                return None
-
-            await _lock_delivery_chain(conn, delivery)
-            if await _has_succeeded_chain_attempt(conn, delivery, delivery_id):
-                await _abandon_current_attempt_after_succeeded_chain_peer(conn, delivery_id)
-                return None
-
-            if delivery["status"] == "retrying" and await _has_successor_attempt(conn, delivery):
-                await conn.execute(
-                    """
-                    UPDATE webhook_deliveries
-                    SET status='abandoned',
-                        superseded=TRUE,
-                        lease_token=NULL,
-                        lease_expires_at=NULL
-                    WHERE id=$1::uuid AND status='retrying' AND NOT superseded
-                    """,
-                    delivery_id,
-                )
-                return None
-
-            pre_claim_monotonic = time.monotonic()
-            claimed = await conn.fetchrow(
-                """
-                UPDATE webhook_deliveries d
-                SET lease_token=$2::uuid,
-                    lease_expires_at=claim_clock.claim_now + ($3::int * INTERVAL '1 second'),
-                    status=CASE WHEN d.status = 'pending' THEN 'retrying' ELSE d.status END
-                FROM webhook_subscriptions s,
-                     (SELECT clock_timestamp() AS claim_now) claim_clock
-                WHERE s.id = d.subscription_id
-                  AND d.id=$1::uuid
-                  AND d.scheduled_at <= claim_clock.claim_now
-                  AND d.attempt_num <= $4
-                  AND NOT d.superseded
-                  AND d.status IN ('pending', 'retrying')
-                  AND (d.lease_token IS NULL OR d.lease_expires_at < claim_clock.claim_now)
-                  AND d.writer_revision = $5
-                  AND (
-                    d.status = 'pending'
-                    OR (
-                      d.status = 'retrying'
-                      AND NOT EXISTS (
-                        SELECT 1
-                        FROM webhook_deliveries newer
-                        WHERE newer.subscription_id = d.subscription_id
-                          AND newer.event_type = d.event_type
-                          AND newer.payload_hash = d.payload_hash
-                          AND newer.attempt_num > d.attempt_num
-                      )
-                    )
-                  )
-                RETURNING d.id, d.subscription_id, d.event_type, d.payload,
-                          d.payload_hash, d.attempt_num, d.status,
-                          d.lease_expires_at, claim_clock.claim_now AS claim_db_now,
-                          s.url, s.secret, s.revoked
-                """,
-                delivery_id,
-                lease_token,
-                lease_seconds,
-                MAX_ATTEMPTS,
-                NEW_CODE_WRITER_REVISION,
-            )
-            if claimed is None:
-                return None
-            return _ClaimedDelivery(
-                delivery=claimed,
-                lease_token=lease_token,
-                pre_claim_monotonic=pre_claim_monotonic,
-            )
-
-
-async def _guard_preclaimed_delivery_before_send(
-    pool: asyncpg.Pool,
-    delivery: asyncpg.Record,
-    lease_token: str,
-) -> bool:
-    """Re-check a recovery-preclaimed row under the chain lock before POSTing."""
-    delivery_id = str(delivery["id"])
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            await _lock_delivery_chain(conn, delivery)
-            if not await _preclaimed_delivery_is_live_and_owned(conn, delivery_id, lease_token):
-                await _release_owned_lease_for_reclaim(conn, delivery_id, lease_token)
-                return False
-
-            if await _has_succeeded_chain_attempt(conn, delivery, delivery_id):
-                await _abandon_owned_attempt_after_succeeded_chain_peer(
-                    conn,
-                    delivery_id,
-                    lease_token,
-                    _DeliveryResult(
-                        succeeded=False,
-                        error="succeeded-chain-peer-before-send",
-                    ),
-                )
-                return False
-
-            if await _has_live_successor_attempt(conn, delivery):
-                await _abandon_owned_attempt_after_live_successor(
-                    conn,
-                    delivery_id,
-                    lease_token,
-                )
-                return False
-
-            return True
-
-
-async def _preclaimed_delivery_is_live_and_owned(
-    conn: asyncpg.Connection,
-    delivery_id: str,
-    lease_token: str,
-) -> bool:
-    """Return whether a preclaimed row is still live under this worker's lease."""
-    return await conn.fetchval(
-        """
-        SELECT EXISTS (
-          SELECT 1
-          FROM webhook_deliveries
-          WHERE id=$1::uuid
-            AND lease_token=$2::uuid
-            AND lease_expires_at > clock_timestamp()
-            AND status IN ('pending', 'retrying')
-            AND NOT superseded
-        )
-        """,
-        delivery_id,
-        lease_token,
-    )
-
-
-async def _send_claimed_delivery(
-    delivery: asyncpg.Record,
-    *,
-    pre_claim_monotonic: float,
-) -> _DeliveryResult:
-    """Perform DNS validation and HTTP POST for an already leased row."""
-    remaining_seconds = _claim_remaining_send_window_seconds(
-        delivery,
-        pre_claim_monotonic=pre_claim_monotonic,
-    )
-    if remaining_seconds <= MIN_SEND_WINDOW_SECONDS:
-        return _LeaseExpiredBeforeSend(
-            succeeded=False,
-            error=(
-                "lease-expired-before-send: remaining lease send window "
-                f"{remaining_seconds:.3f}s <= {MIN_SEND_WINDOW_SECONDS:.1f}s minimum"
-            ),
-        )
-
-    try:
-        return await _send_claimed_delivery_within_deadline(
-            delivery,
-            send_window_seconds=remaining_seconds,
-        )
-    except asyncio.TimeoutError:
-        return _DeliveryResult(
-            succeeded=False,
-            error=(
-                "send-timeout: DNS validation, HTTP send, and response headers "
-                f"exceeded {remaining_seconds:.1f}s lease-anchored wall-clock deadline"
-            ),
-        )
-
-
-async def _send_claimed_delivery_within_deadline(
-    delivery: asyncpg.Record,
-    *,
-    send_window_seconds: float,
-) -> _DeliveryResult:
-    """Run the network send path; caller supplies the wall-clock deadline."""
-    if not delivery:
-        return _DeliveryResult(succeeded=False, error="delivery not found")
-    if delivery["revoked"]:
-        return _DeliveryResult(succeeded=False, error="subscription revoked")
-
-    signature = _sign(delivery["secret"], delivery["payload"])
-    headers = {
-        "Content-Type": "application/json",
-        "User-Agent": "MNEMOS-Webhook/1.0",
-        "Accept-Encoding": "identity",
-        "X-MNEMOS-Event": delivery["event_type"],
-        "X-MNEMOS-Signature": f"sha256={signature}",
-        "X-MNEMOS-Delivery-ID": str(delivery["id"]),
-        "X-MNEMOS-Subscription-ID": str(delivery["subscription_id"]),
-        "X-MNEMOS-Attempt": str(delivery["attempt_num"]),
-    }
-
-    loop = asyncio.get_running_loop()
-    header_deadline = loop.time() + send_window_seconds
-
-    try:
-        from mnemos.api.routes.webhooks import validate_webhook_url
-        await asyncio.wait_for(
-            validate_webhook_url(delivery["url"]),
-            timeout=_remaining_timeout_seconds(header_deadline),
-        )
-    except asyncio.TimeoutError:
-        raise
-    except Exception as e:
-        return _DeliveryResult(succeeded=False, error=f"url-rejected: {type(e).__name__}: {e}")
-
-    client_cm = httpx.AsyncClient(timeout=DELIVERY_TIMEOUT, follow_redirects=False)
-    client = None
-    stream_cm = None
-    try:
-        client = await asyncio.wait_for(
-            client_cm.__aenter__(),
-            timeout=_remaining_timeout_seconds(header_deadline),
-        )
-        stream_cm = client.stream(
-            "POST",
-            delivery["url"],
-            content=delivery["payload"].encode("utf-8"),
-            headers=headers,
-        )
-        response = await asyncio.wait_for(
-            stream_cm.__aenter__(),
-            timeout=_remaining_timeout_seconds(header_deadline),
-        )
-        response_status = response.status_code
-        succeeded = 200 <= response_status < 300
-        return _PostHeaderDeliveryResult(
-            succeeded=succeeded,
-            response_status=response_status,
-            delivery_id=delivery["id"],
-            response=response,
-            stream_cm=stream_cm,
-            client_cm=client_cm,
-        )
-    except asyncio.TimeoutError:
-        await _cleanup_unacknowledged_send_context(
-            delivery_id=delivery["id"],
-            stream_cm=stream_cm,
-            client_cm=client_cm,
-            client=client,
-        )
-        raise
-    except httpx.HTTPError as e:
-        await _cleanup_unacknowledged_send_context(
-            delivery_id=delivery["id"],
-            stream_cm=stream_cm,
-            client_cm=client_cm,
-            client=client,
-        )
-        return _DeliveryResult(succeeded=False, error=f"{type(e).__name__}: {e}")
-    except Exception as e:  # pragma: no cover
-        await _cleanup_unacknowledged_send_context(
-            delivery_id=delivery["id"],
-            stream_cm=stream_cm,
-            client_cm=client_cm,
-            client=client,
-        )
-        return _DeliveryResult(succeeded=False, error=f"{type(e).__name__}: {e}")
-
-
-async def _cleanup_unacknowledged_send_context(
-    *,
-    delivery_id: Any,
-    stream_cm: Any,
-    client_cm: Any,
-    client: Any,
-) -> None:
-    """Best-effort cleanup when the POST did not produce response headers."""
-    if stream_cm is not None:
-        await _run_pre_header_cleanup(
-            stream_cm.__aexit__(None, None, None),
-            delivery_id=delivery_id,
-            cleanup_name="stream",
-        )
-    if client is not None:
-        await _run_pre_header_cleanup(
-            client_cm.__aexit__(None, None, None),
-            delivery_id=delivery_id,
-            cleanup_name="client",
-        )
-
-
-async def _run_pre_header_cleanup(
-    cleanup: Awaitable[object],
-    *,
-    delivery_id: Any,
-    cleanup_name: str,
-) -> None:
-    """Drain failed pre-header send resources without changing the delivery result."""
-    try:
-        await asyncio.wait_for(
-            cleanup,
-            timeout=WEBHOOK_POST_HEADER_CLEANUP_TIMEOUT_SECONDS,
-        )
-    except asyncio.CancelledError:
-        raise
-    except Exception:
-        logger.warning(
-            "webhook delivery %s %s cleanup failed before response headers",
-            delivery_id,
-            cleanup_name,
-            exc_info=True,
-        )
-
-
-async def _run_post_header_cleanup(
-    cleanup: Awaitable[object],
-    *,
-    delivery_id: Any,
-    cleanup_name: str,
-    result: _DeliveryResult | None,
-) -> None:
-    """Bound post-header cleanup so finalization is not stuck behind __aexit__."""
-    cleanup_task = asyncio.ensure_future(cleanup)
-    try:
-        await asyncio.wait_for(
-            asyncio.shield(cleanup_task),
-            timeout=WEBHOOK_POST_HEADER_CLEANUP_TIMEOUT_SECONDS,
-        )
-    except asyncio.TimeoutError:
-        cleanup_task.cancel()
-        cleanup_task.add_done_callback(
-            lambda task: _consume_timed_out_cleanup_result(
-                task,
-                delivery_id,
-                cleanup_name,
-            )
-        )
-        if result is None:
-            raise
-        logger.warning(
-            "webhook delivery %s %s cleanup timed out after %.1fs after response headers; "
-            "delivery result preserved",
-            delivery_id,
-            cleanup_name,
-            WEBHOOK_POST_HEADER_CLEANUP_TIMEOUT_SECONDS,
-            exc_info=True,
-        )
-    except asyncio.CancelledError:
-        cleanup_task.cancel()
-        raise
-    except Exception:
-        if result is None:
-            raise
-        logger.warning(
-            "webhook delivery %s %s cleanup failed after response headers; "
-            "delivery result preserved",
-            delivery_id,
-            cleanup_name,
-            exc_info=True,
-        )
-
-
-def _consume_timed_out_cleanup_result(
-    task: asyncio.Future[object],
-    delivery_id: Any,
-    cleanup_name: str,
-) -> None:
-    """Drain a timed-out cleanup task if it finishes after finalization moved on."""
-    try:
-        task.result()
-    except asyncio.CancelledError:
-        return
-    except Exception:
-        logger.debug(
-            "webhook delivery %s %s cleanup finished with an error after timeout",
-            delivery_id,
-            cleanup_name,
-            exc_info=True,
-        )
-
-
-def _remaining_timeout_seconds(deadline: float) -> float:
-    """Return a small positive timeout if the header deadline is exhausted."""
-    return max(0.001, deadline - asyncio.get_running_loop().time())
-
-
-async def _capture_response_body_for_audit(
-    response: httpx.Response,
-    *,
-    delivery_id: Any,
-) -> Optional[str]:
-    """Read response body for audit after the status-code result is finalized."""
-    try:
-        async with asyncio.timeout(WEBHOOK_RESPONSE_BODY_CAPTURE_TIMEOUT_SECONDS):
-            return await _read_capped_response_body(response)
-    except asyncio.TimeoutError:
-        logger.warning(
-            "webhook delivery %s response-body audit capture timed out after %.1fs; "
-            "leaving response_body NULL",
-            delivery_id,
-            WEBHOOK_RESPONSE_BODY_CAPTURE_TIMEOUT_SECONDS,
-            exc_info=True,
-        )
-        return None
-    except httpx.HTTPError as e:
-        return f"[body capture: {type(e).__name__}]"
-    except Exception as e:  # pragma: no cover
-        return f"[body capture: {type(e).__name__}]"
-
-
-async def _read_capped_response_body(response: httpx.Response) -> str:
-    """Read bounded raw response bytes; never transparently decompress."""
-    headers = getattr(response, "headers", {})
-    content_encoding = str(headers.get("content-encoding", "identity") or "identity").strip().lower()
-    if content_encoding not in ("", "identity"):
-        # Receivers may ignore Accept-Encoding: identity; retain only raw bytes
-        # so a compressed response cannot inflate before the audit cap applies.
-        raw = await _read_capped_raw_response_body(
-            response,
-            max_bytes=min(WEBHOOK_RESPONSE_BODY_MAX_BYTES, NON_IDENTITY_RESPONSE_BODY_PREVIEW_BYTES),
-        )
-        return _decode_capped_response_body(raw, WEBHOOK_RESPONSE_BODY_MAX_BYTES)
-
-    raw = await _read_capped_raw_response_body(response, max_bytes=WEBHOOK_RESPONSE_BODY_MAX_BYTES)
-    return _decode_capped_response_body(raw, WEBHOOK_RESPONSE_BODY_MAX_BYTES)
-
-
-async def _read_capped_raw_response_body(response: httpx.Response, *, max_bytes: int) -> bytes:
-    """Read at most max_bytes raw bytes from a streamed response."""
-    remaining = max_bytes
-    body = bytearray()
-    async for chunk in response.aiter_raw():
-        if not chunk:
-            continue
-        body.extend(chunk[:remaining])
-        remaining -= min(len(chunk), remaining)
-        if remaining <= 0:
-            break
-    return bytes(body)
-
-
-def _decode_capped_response_body(raw: bytes, max_bytes: int) -> str:
-    """Decode for TEXT storage while preserving the configured UTF-8 byte cap."""
-    text = raw.decode("utf-8", errors="replace")
-    if len(text.encode("utf-8")) <= max_bytes:
-        return text
-
-    used = 0
-    out: list[str] = []
-    for char in text:
-        char_size = len(char.encode("utf-8"))
-        if used + char_size > max_bytes:
-            break
-        out.append(char)
-        used += char_size
-    return "".join(out)
-
-
-def _claim_remaining_send_window_seconds(
-    delivery: asyncpg.Record,
-    *,
-    pre_claim_monotonic: float,
-) -> float:
-    """Compute remaining send time from DB claim timestamps plus app elapsed time."""
-    lease_expires_at = _as_aware_utc(delivery["lease_expires_at"])
-    claim_db_now = _as_aware_utc(delivery["claim_db_now"])
-    db_claim_window = (lease_expires_at - claim_db_now).total_seconds()
-    now_monotonic = time.monotonic()
-    elapsed_since_pre_claim = max(0.0, now_monotonic - pre_claim_monotonic)
-    return db_claim_window - elapsed_since_pre_claim - WEBHOOK_FINALIZE_BUFFER_SECONDS
-
-
-def _as_aware_utc(value: datetime) -> datetime:
-    """Normalize asyncpg timestamp values for arithmetic."""
-    if value.tzinfo is None:
-        return value.replace(tzinfo=timezone.utc)
-    return value.astimezone(timezone.utc)
-
-
-async def _finalize_delivery(
-    pool: asyncpg.Pool,
-    delivery: asyncpg.Record,
-    lease_token: str,
-    result: _DeliveryResult,
-) -> bool:
-    """Finalize the delivery row before post-header body capture or cleanup."""
-    finalized = False
-    try:
-        finalized = await _finalize_delivery_row(pool, delivery, lease_token, result)
-        return finalized
-    finally:
-        if isinstance(result, _PostHeaderDeliveryResult):
-            await _run_post_finalize_delivery_work(
-                pool,
-                result,
-                finalized=finalized,
-            )
-
-
-async def _finalize_delivery_row(
-    pool: asyncpg.Pool,
-    delivery: asyncpg.Record,
-    lease_token: str,
-    result: _DeliveryResult,
-) -> bool:
-    """Persist the send result while preserving the attempt ownership rules."""
-    delivery_id = str(delivery["id"])
-    if isinstance(result, _LeaseExpiredBeforeSend):
-        async with pool.acquire() as conn:
-            async with conn.transaction():
-                return await _release_owned_lease_for_reclaim(conn, delivery_id, lease_token)
-
-    if result.succeeded and not delivery["revoked"]:
-        return await _finalize_successful_delivery_row(
-            pool,
-            delivery,
-            delivery_id,
-            lease_token,
-            result,
-        )
-
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            await _lock_delivery_chain(conn, delivery)
-
-            if delivery["revoked"]:
-                finalized = await conn.fetchrow(
-                    """
-                    UPDATE webhook_deliveries
-                    SET status='abandoned',
-                        superseded=FALSE,
-                        error='subscription revoked',
-                        delivered_at=clock_timestamp(),
-                        lease_token=NULL,
-                        lease_expires_at=NULL
-                    WHERE id=$1::uuid
-                      AND lease_token=$2::uuid
-                      AND lease_expires_at >= clock_timestamp()
-                      AND status IN ('pending', 'retrying')
-                      AND NOT superseded
-                    RETURNING id
-                    """,
-                    delivery_id, lease_token,
-                )
-                return finalized is not None
-
-            if await _has_succeeded_chain_attempt(conn, delivery, delivery_id):
-                finalized = await _abandon_owned_attempt_after_succeeded_chain_peer(
-                    conn,
-                    delivery_id,
-                    lease_token,
-                    result,
-                )
-                return finalized is not None
-
-            next_attempt = delivery["attempt_num"] + 1
-            if next_attempt > MAX_ATTEMPTS:
-                finalized = await conn.fetchrow(
-                    """
-                    UPDATE webhook_deliveries
-                    SET status='abandoned',
-                        superseded=FALSE,
-                        response_status=$3,
-                        response_body=$4,
-                        error=$5,
-                        delivered_at=clock_timestamp(),
-                        lease_token=NULL,
-                        lease_expires_at=NULL
-                    WHERE id=$1::uuid
-                      AND lease_token=$2::uuid
-                      AND lease_expires_at >= clock_timestamp()
-                      AND status IN ('pending', 'retrying')
-                      AND NOT superseded
-                    RETURNING id
-                    """,
-                    delivery_id, lease_token, result.response_status, result.response_body, result.error,
-                )
-                return finalized is not None
-
-            backoff = BACKOFF_SCHEDULE[delivery["attempt_num"] - 1]
-            scheduled_at = datetime.now(timezone.utc) + timedelta(seconds=backoff)
-            successor_exists = await _has_successor_attempt(conn, delivery)
-            finalized = await conn.fetchrow(
-                """
-                UPDATE webhook_deliveries
-                SET status='abandoned',
-                    superseded=TRUE,
-                    response_status=$3,
-                    response_body=$4,
-                    error=$5,
-                    lease_token=NULL,
-                    lease_expires_at=NULL
-                WHERE id=$1::uuid
-                  AND lease_token=$2::uuid
-                  AND lease_expires_at >= clock_timestamp()
-                  AND status IN ('pending', 'retrying')
-                  AND NOT superseded
-                RETURNING id
-                """,
-                delivery_id,
-                lease_token,
-                result.response_status,
-                result.response_body,
-                result.error,
-            )
-            if finalized is None:
-                logger.info(
-                    "webhook delivery %s finalize skipped because lease expired or moved",
-                    delivery_id,
-                )
-                return False
-
-            if not successor_exists:
-                await _insert_successor_delivery(conn, delivery, next_attempt, scheduled_at)
-            logger.info(
-                "webhook delivery %s attempt %d failed (status=%s error=%s), retry in %ds",
-                delivery_id, delivery["attempt_num"], result.response_status, result.error, backoff,
-            )
-            return True
-
-
-async def _commit_successful_delivery_row(
-    pool: asyncpg.Pool,
-    delivery: asyncpg.Record,
-    delivery_id: str,
-    lease_token: str,
-    result: _DeliveryResult,
-) -> bool:
-    """Commit a 2xx result with same-transaction successor convergence."""
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            await _lock_delivery_chain(conn, delivery)
-
-            if await _has_succeeded_chain_attempt(conn, delivery, delivery_id):
-                finalized = await _abandon_owned_attempt_after_succeeded_chain_peer(
-                    conn,
-                    delivery_id,
-                    lease_token,
-                    result,
-                    require_unexpired_lease=False,
-                )
-                return finalized is not None
-
-            finalized = await conn.fetchrow(
-                """
-                UPDATE webhook_deliveries
-                SET status='succeeded',
-                    superseded=FALSE,
-                    response_status=$3,
-                    response_body=$4,
-                    error=NULL,
-                    delivered_at=clock_timestamp(),
-                    lease_token=NULL,
-                    lease_expires_at=NULL
-                WHERE id=$1::uuid
-                  AND lease_token=$2::uuid
-                  AND status IN ('pending', 'retrying')
-                  AND NOT superseded
-                RETURNING id
-                """,
-                delivery_id,
-                lease_token,
-                result.response_status,
-                result.response_body,
-            )
-            if finalized is None:
-                finalized = await _abandon_owned_attempt_after_succeeded_chain_peer(
-                    conn,
-                    delivery_id,
-                    lease_token,
-                    result,
-                    require_unexpired_lease=False,
-                )
-                if finalized is not None:
-                    return True
-                await _clear_stale_owned_lease_after_terminal_finalize(
-                    conn,
-                    delivery_id,
-                    lease_token,
-                )
-                return False
-
-            await _abandon_live_successors_before_success_commit(conn, delivery)
-            return True
-
-
-async def _finalize_successful_delivery_row(
-    pool: asyncpg.Pool,
-    delivery: asyncpg.Record,
-    delivery_id: str,
-    lease_token: str,
-    result: _DeliveryResult,
-) -> bool:
-    """Persist a 2xx ACK with atomic successor convergence."""
-    try:
-        finalized = await _commit_successful_delivery_row(
-            pool,
-            delivery,
-            delivery_id,
-            lease_token,
-            result,
-        )
-    except asyncpg.exceptions.UniqueViolationError:
-        return await _abandon_success_duplicate_after_unique_violation(
-            pool,
-            delivery,
-            delivery_id,
-            lease_token,
-            result,
-        )
-    except asyncio.CancelledError:
-        logger.warning(
-            "webhook delivery %s success finalization cancelled before commit after 2xx ACK; "
-            "success transaction rolled back and retry may resend a bounded duplicate POST",
-            delivery_id,
-            exc_info=True,
-        )
-        raise
-    except Exception:
-        logger.warning(
-            "webhook delivery %s success finalization failed before commit after 2xx ACK; "
-            "success transaction rolled back and retry may resend a bounded duplicate POST",
-            delivery_id,
-            exc_info=True,
-        )
-        raise
-
-    return finalized
-
-
-async def _abandon_live_successors_before_success_commit(
-    conn: asyncpg.Connection,
-    delivery: asyncpg.Record,
-) -> None:
-    """Abandon free successors inside the success transaction."""
-    successors = await _find_live_unleased_successor_attempts(conn, delivery)
-    for successor in successors:
-        await _abandon_live_successor_attempt(conn, str(successor["id"]))
-
-
-async def _abandon_success_duplicate_after_unique_violation(
-    pool: asyncpg.Pool,
-    delivery: asyncpg.Record,
-    delivery_id: str,
-    lease_token: str,
-    result: _DeliveryResult,
-) -> bool:
-    """Converge a duplicate 2xx after the succeeded-chain unique index wins."""
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            await _lock_delivery_chain(conn, delivery)
-            finalized = await _abandon_owned_attempt_after_succeeded_chain_peer(
-                conn,
-                delivery_id,
-                lease_token,
-                result,
-                require_unexpired_lease=False,
-            )
-            if finalized is None:
-                await _clear_stale_owned_lease_after_terminal_finalize(
-                    conn,
-                    delivery_id,
-                    lease_token,
-                )
-                return False
-            return True
-
-
-async def _run_post_finalize_delivery_work(
-    pool: asyncpg.Pool,
-    result: _PostHeaderDeliveryResult,
-    *,
-    finalized: bool,
-) -> None:
-    """Capture audit body and close HTTP resources after the DB result is durable."""
-    try:
-        if finalized:
-            response_body = await _capture_response_body_for_audit(
-                result.response,
-                delivery_id=result.delivery_id,
-            )
-            if response_body is not None:
-                await _persist_response_body_for_audit(
-                    pool,
-                    delivery_id=result.delivery_id,
-                    response_body=response_body,
-                )
-    finally:
-        if result.stream_cm is not None:
-            await _run_post_header_cleanup(
-                result.stream_cm.__aexit__(None, None, None),
-                delivery_id=result.delivery_id,
-                cleanup_name="stream",
-                result=result,
-            )
-        if result.client_cm is not None:
-            await _run_post_header_cleanup(
-                result.client_cm.__aexit__(None, None, None),
-                delivery_id=result.delivery_id,
-                cleanup_name="client",
-                result=result,
-            )
-
-
-async def _persist_response_body_for_audit(
-    pool: asyncpg.Pool,
-    *,
-    delivery_id: Any,
-    response_body: str,
-) -> None:
-    """Store post-finalize response audit data without lease ownership checks."""
-    try:
-        async with pool.acquire() as conn:
-            await conn.execute(
-                """
-                UPDATE webhook_deliveries
-                SET response_body=$2
-                WHERE id=$1::uuid
-                """,
-                str(delivery_id),
-                response_body,
-            )
-    except asyncio.CancelledError:
-        raise
-    except Exception:
-        logger.warning(
-            "webhook delivery %s response-body audit update failed after finalization",
-            delivery_id,
-            exc_info=True,
-        )
-
-
-async def _release_owned_lease_for_reclaim(
-    conn: asyncpg.Connection,
-    delivery_id: str,
-    lease_token: str,
-) -> bool:
-    """Clear this worker's lease after no POST occurred, without consuming a retry."""
-    released = await conn.fetchrow(
-        """
-        UPDATE webhook_deliveries
-        SET lease_token=NULL,
-            lease_expires_at=NULL
-        WHERE id=$1::uuid
-          AND lease_token=$2::uuid
-        RETURNING id, status, superseded
-        """,
-        delivery_id,
-        lease_token,
-    )
-    return (
-        released is not None
-        and released["status"] in LIVE_DELIVERY_STATUSES
-        and not released["superseded"]
-    )
-
-
-async def _clear_stale_owned_lease_after_terminal_finalize(
-    conn: asyncpg.Connection,
-    delivery_id: str,
-    lease_token: str,
-) -> None:
-    """Release our lease when another writer terminalized the same row first."""
-    cleared = await conn.fetchrow(
-        """
-        UPDATE webhook_deliveries
-        SET lease_token=NULL,
-            lease_expires_at=NULL
-        WHERE id=$1::uuid
-          AND lease_token=$2::uuid
-        RETURNING id, status, superseded
-        """,
-        delivery_id,
-        lease_token,
-    )
-    if cleared is None:
-        logger.warning(
-            "webhook delivery %s success finalize found no live owned row; stale lease was already gone",
-            delivery_id,
-        )
-        return
-    if cleared["status"] not in LIVE_DELIVERY_STATUSES or cleared["superseded"]:
-        logger.warning(
-            "webhook delivery %s was already terminal at success finalize time; stale lease cleared",
-            delivery_id,
-        )
-        return
-    logger.warning(
-        "webhook delivery %s success finalize found no live owned row; stale lease cleared",
-        delivery_id,
-    )
-
-
-async def _load_delivery_for_claim(conn: asyncpg.Connection, delivery_id: str) -> Optional[asyncpg.Record]:
-    """Load a due live delivery candidate for a short lease attempt."""
-    return await conn.fetchrow(
-        """
-        SELECT d.id, d.subscription_id, d.event_type, d.payload,
-               d.payload_hash, d.attempt_num, d.status,
-               s.url, s.secret, s.revoked
-        FROM webhook_deliveries d
-        JOIN webhook_subscriptions s ON s.id = d.subscription_id
-        WHERE d.id = $1::uuid
-          AND d.scheduled_at <= clock_timestamp()
-          AND d.attempt_num <= $2
-          AND NOT d.superseded
-          AND d.status IN ('pending', 'retrying')
-          AND (d.lease_token IS NULL OR d.lease_expires_at < clock_timestamp())
-        """,
-        delivery_id, MAX_ATTEMPTS,
-    )
-
-
-async def _insert_successor_delivery(
-    conn: asyncpg.Connection,
-    delivery: asyncpg.Record,
-    next_attempt: int,
-    scheduled_at: datetime,
-) -> Optional[asyncpg.Record]:
-    """Insert the next live retry attempt if another writer has not already won."""
-    return await conn.fetchrow(
-        """
-        INSERT INTO webhook_deliveries
-          (subscription_id, event_type, payload, payload_hash,
-           attempt_num, status, scheduled_at, writer_revision)
-        VALUES ($1, $2, $3, $4, $5, 'pending', $6, $7)
-        ON CONFLICT (subscription_id, event_type, payload_hash, attempt_num)
-          WHERE status IN ('pending', 'retrying') AND NOT superseded
-        DO NOTHING
-        RETURNING id
-        """,
-        delivery["subscription_id"],
-        delivery["event_type"],
-        delivery["payload"],
-        delivery["payload_hash"],
-        next_attempt,
-        scheduled_at,
-        NEW_CODE_WRITER_REVISION,
-    )
-
-
-async def _lock_delivery_chain(conn: asyncpg.Connection, delivery: asyncpg.Record) -> None:
-    """Serialize new-code recovery claims and successor inserts per chain."""
-    await conn.execute("SELECT pg_advisory_xact_lock($1)", _delivery_chain_lock_key(delivery))
-
-
-def _delivery_chain_lock_key(delivery: asyncpg.Record, _hashlib_mod=None) -> int:
-    """Stable signed-int64 lock key for one webhook retry chain."""
-    if _hashlib_mod is None:
-        _hashlib_mod = hashlib
-    digest = _hashlib_mod.sha256(
-        (
-            "webhook-chain:"
-            f"{delivery['subscription_id']}:{delivery['event_type']}:{delivery['payload_hash']}"
-        ).encode("utf-8")
-    ).digest()[:8]
-    key = int.from_bytes(digest, "big", signed=False)
-    if key >= 2**63:
-        key -= 2**64
-    return key
-
-
-async def _has_successor_attempt(conn: asyncpg.Connection, delivery: asyncpg.Record) -> bool:
-    """Return whether a newer attempt already exists for this delivery chain."""
-    return await conn.fetchval(
-        """
-        SELECT EXISTS (
-          SELECT 1
-          FROM webhook_deliveries newer
-          WHERE newer.subscription_id = $1
-            AND newer.event_type = $2
-            AND newer.payload_hash = $3
-            AND newer.attempt_num > $4
-        )
-        """,
-        delivery["subscription_id"],
-        delivery["event_type"],
-        delivery["payload_hash"],
-        delivery["attempt_num"],
-    )
-
-
-async def _has_live_successor_attempt(conn: asyncpg.Connection, delivery: asyncpg.Record) -> bool:
-    """Return whether a live newer attempt owns the chain's forward direction."""
-    return await conn.fetchval(
-        """
-        SELECT EXISTS (
-          SELECT 1
-          FROM webhook_deliveries newer
-          WHERE newer.subscription_id = $1
-            AND newer.event_type = $2
-            AND newer.payload_hash = $3
-            AND newer.attempt_num > $4
-            AND newer.status IN ('pending', 'retrying')
-            AND NOT newer.superseded
-        )
-        """,
-        delivery["subscription_id"],
-        delivery["event_type"],
-        delivery["payload_hash"],
-        delivery["attempt_num"],
-    )
-
-
-async def _has_succeeded_chain_attempt(
-    conn: asyncpg.Connection,
-    delivery: asyncpg.Record,
-    delivery_id: str,
-) -> bool:
-    """Return whether any attempt already completed the chain."""
-    return await conn.fetchval(
-        """
-        SELECT EXISTS (
-          SELECT 1
-          FROM webhook_deliveries peer
-          WHERE peer.subscription_id = $1
-            AND peer.event_type = $2
-            AND peer.payload_hash = $3
-            AND peer.status = 'succeeded'
-            AND peer.id <> $4::uuid
-        )
-        """,
-        delivery["subscription_id"],
-        delivery["event_type"],
-        delivery["payload_hash"],
-        delivery_id,
-    )
-
-
-async def _abandon_owned_attempt_after_live_successor(
-    conn: asyncpg.Connection,
-    delivery_id: str,
-    lease_token: str,
-) -> Optional[asyncpg.Record]:
-    """Terminalize an owned preclaimed attempt made obsolete before POST."""
-    return await conn.fetchrow(
-        """
-        UPDATE webhook_deliveries
-        SET status='abandoned',
-            superseded=TRUE,
-            status_updated_at=clock_timestamp(),
-            lease_token=NULL,
-            lease_expires_at=NULL
-        WHERE id=$1::uuid
-          AND lease_token=$2::uuid
-          AND lease_expires_at > clock_timestamp()
-          AND status IN ('pending', 'retrying')
-          AND NOT superseded
-        RETURNING id
-        """,
-        delivery_id,
-        lease_token,
-    )
-
-
-async def _abandon_current_attempt_after_succeeded_chain_peer(
-    conn: asyncpg.Connection,
-    delivery_id: str,
-) -> None:
-    """Terminalize a lease-free attempt whose chain already converged."""
-    await conn.execute(
-        """
-        UPDATE webhook_deliveries
-        SET status='abandoned',
-            superseded=TRUE,
-            status_updated_at=clock_timestamp(),
-            lease_token=NULL,
-            lease_expires_at=NULL
-        WHERE id=$1::uuid
-          AND status IN ('pending', 'retrying')
-          AND NOT superseded
-          AND (lease_token IS NULL OR lease_expires_at < clock_timestamp())
-        """,
-        delivery_id,
-    )
-
-
-async def _abandon_owned_attempt_after_succeeded_chain_peer(
-    conn: asyncpg.Connection,
-    delivery_id: str,
-    lease_token: str,
-    result: _DeliveryResult,
-    *,
-    require_unexpired_lease: bool = True,
-) -> Optional[asyncpg.Record]:
-    """Finalize an active duplicate without extending a completed chain."""
-    if require_unexpired_lease:
-        return await conn.fetchrow(
-            """
-            UPDATE webhook_deliveries
-            SET status='abandoned',
-                superseded=TRUE,
-                response_status=$3,
-                response_body=$4,
-                error=$5,
-                lease_token=NULL,
-                lease_expires_at=NULL
-            WHERE id=$1::uuid
-              AND lease_token=$2::uuid
-              AND lease_expires_at >= clock_timestamp()
-              AND status IN ('pending', 'retrying')
-              AND NOT superseded
-            RETURNING id
-            """,
-            delivery_id,
-            lease_token,
-            result.response_status,
-            result.response_body,
-            result.error,
-        )
-
-    return await conn.fetchrow(
-        """
-        UPDATE webhook_deliveries
-        SET status='abandoned',
-            superseded=TRUE,
-            response_status=$3,
-            response_body=$4,
-            error=$5,
-            lease_token=NULL,
-            lease_expires_at=NULL
-        WHERE id=$1::uuid
-          AND lease_token=$2::uuid
-          AND status IN ('pending', 'retrying')
-          AND NOT superseded
-        RETURNING id
-        """,
-        delivery_id,
-        lease_token,
-        result.response_status,
-        result.response_body,
-        result.error,
-    )
-
-
-async def _find_live_unleased_successor_attempts(
-    conn: asyncpg.Connection,
-    delivery: asyncpg.Record,
-) -> list[asyncpg.Record]:
-    """Return all free live successors that would duplicate this success."""
-    return await conn.fetch(
-        """
-        SELECT newer.id
-        FROM webhook_deliveries newer
-        WHERE newer.subscription_id = $1
-          AND newer.event_type = $2
-          AND newer.payload_hash = $3
-          AND newer.attempt_num > $4
-          AND newer.status IN ('pending', 'retrying')
-          AND NOT newer.superseded
-          AND (newer.lease_token IS NULL OR newer.lease_expires_at < clock_timestamp())
-        ORDER BY newer.attempt_num ASC
-        """,
-        delivery["subscription_id"],
-        delivery["event_type"],
-        delivery["payload_hash"],
-        delivery["attempt_num"],
-    )
-
-
-async def _abandon_live_successor_attempt(conn: asyncpg.Connection, successor_id: str) -> None:
-    """Mark a free successor superseded after a predecessor has succeeded."""
-    await conn.execute(
-        """
-        UPDATE webhook_deliveries
-        SET status='abandoned',
-            superseded=TRUE,
-            status_updated_at=clock_timestamp(),
-            lease_token=NULL,
-            lease_expires_at=NULL
-        WHERE id=$1::uuid
-          AND status IN ('pending', 'retrying')
-          AND NOT superseded
-          AND (lease_token IS NULL OR lease_expires_at < clock_timestamp())
-        """,
-        successor_id,
-    )
-
-
-def _sign(secret: str, body: str) -> str:
-    return hmac.new(
-        secret.encode("utf-8"),
-        body.encode("utf-8"),
-        hashlib.sha256,
-    ).hexdigest()
+# Legacy private attributes are resolved lazily so older in-repo tests and
+# debugging imports can still reach the moved implementation without putting
+# state-machine code back in this public dispatcher module.
+_LEGACY_MODULES = {
+    "asyncio": asyncio,
+    "time": time,
+    "asyncpg": asyncpg,
+    "httpx": httpx,
+}
+_LEGACY_ATTR_TARGETS: dict[str, object] = {}
+
+
+def _register_legacy_attrs(target: object, names: tuple[str, ...]) -> None:
+    for name in names:
+        _LEGACY_ATTR_TARGETS[name] = target
+
+
+_register_legacy_attrs(webhook_types, (
+    "BACKOFF_SCHEDULE",
+    "MAX_ATTEMPTS",
+    "DNS_TIMEOUT",
+    "DELIVERY_TIMEOUT",
+    "WEBHOOK_LEASE_SECONDS",
+    "WEBHOOK_FINALIZE_BUFFER_SECONDS",
+    "WEBHOOK_RESPONSE_BODY_MAX_BYTES",
+    "WEBHOOK_RESPONSE_BODY_CAPTURE_TIMEOUT_SECONDS",
+    "WEBHOOK_POST_HEADER_CLEANUP_TIMEOUT_SECONDS",
+    "WEBHOOK_MAX_CONCURRENT_SENDS",
+    "NEW_CODE_WRITER_REVISION",
+    "NON_IDENTITY_RESPONSE_BODY_PREVIEW_BYTES",
+    "MIN_SEND_WINDOW_SECONDS",
+    "RECOVERY_POLL_INTERVAL",
+    "REPAIR_BURST_SECONDS",
+    "REPAIR_BURST_INTERVAL",
+    "REPAIR_PERIODIC_INTERVAL",
+    "TERMINAL_DELIVERY_STATUSES",
+    "LIVE_DELIVERY_STATUSES",
+    "TOTAL_SEND_DEADLINE_SECONDS",
+    "_send_semaphore",
+    "_DeliveryResult",
+    "_LeaseExpiredBeforeSend",
+    "_PostHeaderDeliveryResult",
+    "_ClaimedDelivery",
+    "_derive_total_send_deadline_seconds",
+    "_get_send_semaphore",
+))
+_register_legacy_attrs(webhook_signing, ("_sign",))
+_register_legacy_attrs(webhook_workers, (
+    "repair_worker_loop",
+    "delivery_worker_loop",
+    "recovery_worker_loop",
+    "_recover_due_deliveries",
+    "_semaphore_available",
+    "_claim_recoverable_deliveries",
+    "_recoverable_delivery_ids",
+))
+_register_legacy_attrs(webhook_repair, (
+    "_repair_superseded_retrying_deliveries_safely",
+    "repair_superseded_retrying_deliveries",
+))
+_register_legacy_attrs(webhook_sender, (
+    "_attempt_delivery",
+    "_send_claimed_delivery",
+    "_send_claimed_delivery_within_deadline",
+    "_cleanup_unacknowledged_send_context",
+    "_run_pre_header_cleanup",
+    "_run_post_header_cleanup",
+    "_consume_timed_out_cleanup_result",
+    "_remaining_timeout_seconds",
+    "_capture_response_body_for_audit",
+    "_read_capped_response_body",
+    "_read_capped_raw_response_body",
+    "_decode_capped_response_body",
+))
+_register_legacy_attrs(webhook_lease, (
+    "_claim_delivery",
+    "_guard_preclaimed_delivery_before_send",
+    "_preclaimed_delivery_is_live_and_owned",
+    "_claim_remaining_send_window_seconds",
+    "_as_aware_utc",
+    "_release_owned_lease_for_reclaim",
+    "_clear_stale_owned_lease_after_terminal_finalize",
+))
+_register_legacy_attrs(webhook_finalize, (
+    "_finalize_delivery",
+    "_finalize_delivery_row",
+    "_commit_successful_delivery_row",
+    "_finalize_successful_delivery_row",
+    "_abandon_live_successors_before_success_commit",
+    "_abandon_success_duplicate_after_unique_violation",
+    "_run_post_finalize_delivery_work",
+    "_persist_response_body_for_audit",
+))
+_register_legacy_attrs(webhook_chain, (
+    "_load_delivery_for_claim",
+    "_insert_successor_delivery",
+    "_lock_delivery_chain",
+    "_delivery_chain_lock_key",
+    "_has_successor_attempt",
+    "_has_live_successor_attempt",
+    "_has_succeeded_chain_attempt",
+    "_abandon_owned_attempt_after_live_successor",
+    "_abandon_current_attempt_after_succeeded_chain_peer",
+    "_abandon_owned_attempt_after_succeeded_chain_peer",
+    "_find_live_unleased_successor_attempts",
+    "_abandon_live_successor_attempt",
+))
+
+
+def __getattr__(name: str) -> Any:
+    if name in _LEGACY_MODULES:
+        return _LEGACY_MODULES[name]
+    target = _LEGACY_ATTR_TARGETS.get(name)
+    if target is not None:
+        return getattr(target, name)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+
+class _DispatcherModule(_module_types.ModuleType):
+    def __setattr__(self, name: str, value: Any) -> None:
+        target = _LEGACY_ATTR_TARGETS.get(name)
+        if target is not None:
+            setattr(target, name, value)
+        super().__setattr__(name, value)
+
+
+sys.modules[__name__].__class__ = _DispatcherModule
+
+
+__all__ = [
+    "dispatch",
+    "_dispatch_on_conn",
+]
